Move beam scorer and beam state into a header file
